Refactored some complexity into helper functions for arithmetic
