+++ src/Samsara/Fermat/Types/Traits/Arithmetic/ArithmeticScaleTrait.php
@@ -103,7 +103,7 @@ trait ArithmeticScaleTrait
 
         $scale = $scale ?? $this->getScale();
 
-        return ArithmeticProvider::squareRoot($this->getAsBaseTenRealNumber(), $scale);
+        return ArithmeticProvider::squareRoot($this->abs()->getAsBaseTenRealNumber(), $scale);
 
     }
