+++ src/Samsara/Fermat/Types/Traits/Arithmetic/ArithmeticSelectionTrait.php
@@ -1,179 +1,20 @@
-<?php /** @noinspection ALL */
-/** @noinspection ALL */
-/** @noinspection ALL */
-/** @noinspection ALL */
-/** @noinspection ALL */
-/** @noinspection ALL */
-/** @noinspection ALL */
-/** @noinspection ALL */
-/** @noinspection ALL */
-/** @noinspection ALL */
-/** @noinspection ALL */
-/** @noinspection ALL */
-/** @noinspection ALL */
-/** @noinspection ALL */
-/** @noinspection ALL */
-
-/** @noinspection ALL */
-
+<?php
 
 namespace Samsara\Fermat\Types\Traits\Arithmetic;
 
-
-use Composer\InstalledVersions;
-use Samsara\Exceptions\SystemError\PlatformError\MissingPackage;
 use Samsara\Exceptions\UsageError\IntegrityConstraint;
-use Samsara\Fermat\ComplexNumbers;
 use Samsara\Fermat\Enums\CalcMode;
 use Samsara\Fermat\Enums\CalcOperation;
-use Samsara\Fermat\Enums\NumberBase;
 use Samsara\Fermat\Numbers;
 use Samsara\Fermat\Provider\CalculationModeProvider;
-use Samsara\Fermat\Provider\RoundingProvider;
-use Samsara\Fermat\Types\Base\Interfaces\Coordinates\CoordinateInterface;
-use Samsara\Fermat\Types\Base\Interfaces\Numbers\ComplexNumberInterface;
 use Samsara\Fermat\Types\Base\Interfaces\Numbers\DecimalInterface;
-use Samsara\Fermat\Types\Base\Interfaces\Numbers\FractionInterface;
-use Samsara\Fermat\Types\Base\Interfaces\Numbers\NumberInterface;
-use Samsara\Fermat\Types\ComplexNumber;
-use Samsara\Fermat\Values\Geometry\CoordinateSystems\CartesianCoordinate;
 use Samsara\Fermat\Values\ImmutableDecimal;
-use Samsara\Fermat\Values\ImmutableFraction;
-use Samsara\Fermat\Values\MutableDecimal;
-use Samsara\Fermat\Values\MutableFraction;
 
 /**
  *
  */
 trait ArithmeticSelectionTrait
 {
-
-    /**
-     * @param $left
-     * @param $right
-     * @param int $identity
-     *
-     * @return array
-     * @throws IntegrityConstraint
-     */
-    protected function translateToParts($left, $right, int $identity = 0): array
-    {
-        switch (gettype($right)) {
-            case 'integer':
-            case 'double':
-                $right = Numbers::make(Numbers::IMMUTABLE, $right);
-                break;
-
-            case 'string':
-                $right = self::stringSelector($right);
-                break;
-
-            case 'object':
-                if ($right instanceof MutableDecimal) {
-                    $right = Numbers::makeOrDont(Numbers::IMMUTABLE, $right);
-                } elseif ($right instanceof MutableFraction) {
-                    $right = Numbers::makeOrDont(Numbers::IMMUTABLE_FRACTION, $right);
-                } else {
-                    $right = !($right instanceof NumberInterface) ? Numbers::makeOrDont(Numbers::IMMUTABLE, $right) : $right;
-                }
-                break;
-        }
-
-        [$thatRealPart, $thatImaginaryPart, $right] = self::rightSelector($left, $right, $identity);
-
-        [$thisRealPart, $thisImaginaryPart] = self::leftSelector($left, $identity);
-
-        return [$thatRealPart, $thatImaginaryPart, $thisRealPart, $thisImaginaryPart, $right];
-    }
-
-    /**
-     * @param string $input
-     * @return CoordinateInterface|FractionInterface|NumberInterface|ComplexNumber|CartesianCoordinate|ImmutableDecimal|ImmutableFraction|MutableDecimal|MutableFraction
-     * @throws IntegrityConstraint|MissingPackage
-     */
-    protected static function stringSelector(string $input)
-    {
-
-        $input = trim($input);
-        if (str_contains($input, '/')) {
-            $input = Numbers::makeFractionFromString(Numbers::IMMUTABLE_FRACTION, $input);
-        } elseif (strrpos($input, '+') || strrpos($input, '-')) {
-            if (!(InstalledVersions::isInstalled('samsara/fermat-complex-numbers'))) {
-                throw new MissingPackage(
-                    'Creating complex numbers is unsupported in Fermat without modules.',
-                    'Install the samsara/fermat-complex-numbers package using composer.',
-                    'An attempt was made to create a ComplexNumber instance without having the Complex Numbers module. Please install the samsara/fermat-complex-numbers package using composer.'
-                );
-            }
-
-            $input = ComplexNumbers::make(ComplexNumbers::IMMUTABLE_COMPLEX, $input);
-        } else {
-            $input = Numbers::make(Numbers::IMMUTABLE, $input);
-        }
-
-        return $input;
-
-    }
-
-    /**
-     * @param $left
-     * @param $right
-     * @param $identity
-     * @return array
-     * @throws IntegrityConstraint
-     */
-    protected static function rightSelector($left, $right, $identity): array
-    {
-
-        if ($right instanceof ComplexNumberInterface) {
-            $thatRealPart = $right->getRealPart();
-            $thatImaginaryPart = $right->getImaginaryPart();
-        } else {
-            if ($right instanceof FractionInterface) {
-                if ($left instanceof FractionInterface) {
-                    $rightPart = $right;
-
-                    $thatRealPart = $right->isReal() ? $rightPart : new ImmutableFraction(new ImmutableDecimal($identity), new ImmutableDecimal(1));
-                    $thatImaginaryPart = $right->isImaginary() ? $rightPart : new ImmutableFraction(new ImmutableDecimal($identity.'i'), new ImmutableDecimal(1));
-                } else {
-                    $rightPart = $right->asDecimal();
-                    $right = $right->asDecimal();
-
-                    $thatRealPart = $right->isReal() ? $rightPart : new ImmutableDecimal($identity, $left->getScale());
-                    $thatImaginaryPart = $right->isImaginary() ? $rightPart : new ImmutableDecimal($identity.'i', $left->getScale());
-                }
-            } else {
-                $rightPart = $right;
-
-                $thatRealPart = $right->isReal() ? $rightPart : new ImmutableDecimal($identity, $left->getScale());
-                $thatImaginaryPart = $right->isImaginary() ? $rightPart : new ImmutableDecimal($identity.'i', $left->getScale());
-            }
-        }
-
-        return [$thatRealPart, $thatImaginaryPart, $right];
-
-    }
-
-    /**
-     * @param $left
-     * @param $identity
-     * @return array
-     * @throws IntegrityConstraint
-     */
-    protected static function leftSelector($left, $identity): array
-    {
-
-        if ($left instanceof ComplexNumberInterface) {
-            $thisRealPart = $left->getRealPart();
-            $thisImaginaryPart = $left->getImaginaryPart();
-        } else {
-            $thisRealPart = $left->isReal() ? $left : new ImmutableDecimal($identity, $left->getScale());
-            $thisImaginaryPart = $left->isImaginary() ? $left : new ImmutableDecimal($identity.'i', $left->getScale());
-        }
-
-        return [$thisRealPart, $thisImaginaryPart];
-
-    }
 
     /**
      * @param DecimalInterface $num
@@ -345,6 +186,7 @@ trait ArithmeticSelectionTrait
     protected function sqrtSelector(?int $scale): string
     {
         $calcMode = $this->getMode();
+
         if ($calcMode == CalcMode::Auto) {
             $value = $this->sqrtGMP();
